backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
requests==2.31.0
Pillow>=10.2.0
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0
django-allauth==0.61.1
python-dotenv==1.0.1
